refactored :multiselect_communicator_interface
