+++ libp2p/protocol_muxer/multiselect_client.py
@@ -4,6 +4,7 @@ from collections.abc import (
 
 from libp2p.abc import (
     IMultiselectClient,
+    IMultiselectCommunicator,
 )
 from libp2p.custom_types import (
     TProtocol,
@@ -12,9 +13,6 @@ from libp2p.custom_types import (
 from .exceptions import (
     MultiselectClientError,
     MultiselectCommunicatorError,
-)
-from .multiselect_communicator_interface import (
-    IMultiselectCommunicator,
 )
 
 MULTISELECT_PROTOCOL_ID = "/multistream/1.0.0"
